Split torch module code into torch_layers file
