+++ stable_baselines3/sac/policies.py
@@ -5,9 +5,8 @@ import torch as th
 import torch.nn as nn
 
 from stable_baselines3.common.preprocessing import get_action_dim
-from stable_baselines3.common.policies import (BasePolicy, register_policy, create_mlp,
-                                               create_sde_features_extractor, NatureCNN,
-                                               BaseFeaturesExtractor, FlattenExtractor)
+from stable_baselines3.common.policies import BasePolicy, register_policy, create_sde_features_extractor
+from stable_baselines3.common.torch_layers import create_mlp, NatureCNN, BaseFeaturesExtractor, FlattenExtractor
 from stable_baselines3.common.distributions import SquashedDiagGaussianDistribution, StateDependentNoiseDistribution
 
 # CAP the standard deviation of the actor
